verified and remade some beans, config bean works

+++ src/main/java/spz/dae24/dtos/ClientDTO.java
@@ -1,16 +1,13 @@
 package spz.dae24.dtos;
 
-import jakarta.validation.constraints.NotBlank;
 import spz.dae24.entities.Client;
 import spz.dae24.entities.Package;
-import spz.dae24.entities.Sensor;
-import spz.dae24.entities.Volume;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class ClientDTO {
-    private int id;
+    private long id;
     private String username;
     private String name;
     private String email;
@@ -18,7 +15,7 @@ public class ClientDTO {
 
     public ClientDTO() {}
 
-    public ClientDTO(List<Package> packages, String email, String name, String username, int id) {
+    public ClientDTO(List<Package> packages, String email, String name, String username, long id) {
         this.packages = packages;
         this.email = email;
         this.name = name;
@@ -26,11 +23,11 @@ public class ClientDTO {
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
