+++ src/main/java/spz/dae24/entities/Admin.java
@@ -9,7 +9,7 @@ public class Admin extends User {
 
     public Admin() {}
 
-    public Admin(int id, String username, String name, String email, String password) {
+    public Admin(long id, String username, String name, String email, String password) {
         super(id, username, name, email,password);
     }
 }
